queue component decomposition
