+++ client/src/components/ItemList.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Item from "./Item";
+
+const ItemList = ({queue}) => {
+    return (
+        <div>
+            {queue.map((item) =>
+                <Item item={item}/>
+            )}
+        </div>
+    );
+};
+
+export default ItemList;
