+++ client/src/components/Queue.jsx
@@ -1,16 +1,15 @@
 import React, {useContext, useEffect, useState} from 'react';
 import {over} from 'stompjs';
 import SockJS from 'sockjs-client';
-import Item from "./Item";
 import {UserContext} from "../context";
-import {useNavigate} from "react-router-dom";
+import ItemList from "./ItemList";
+import Navbar from "./Navbar";
 
 const Queue = (props) => {
     let sock = new SockJS(props.url + props.socketEndpoint)
     let stompClient = over(sock)
     const [queue, setQueue] = useState([])
     const {user, setUser} = useContext(UserContext)
-    const navigate  = useNavigate()
     useEffect(() =>{
         stompClient.connect({}, onConnect)
 
@@ -19,6 +18,9 @@ const Queue = (props) => {
             .then(data => {
                 setQueue(data)
             })
+        return () => {
+            sock.close()
+        }
     },[])
     const onConnect = ()=>{
         stompClient.subscribe('/client/take', take)
@@ -60,47 +62,20 @@ const Queue = (props) => {
             })
             return newQueue
         })
-
-        setUser(oldUser => {
-            return {name:oldUser.name, number:0}
-        })
-    }
-    const takeQueue =  () => {
-        stompClient.send("/server/take", {}, JSON.stringify({'name': user.name}))
-    }
-    const leaveQueue = () => {
-        stompClient.send("/server/leave", {}, JSON.stringify({'name': user.name, 'number': user.number}))
-    }
-    const logOut =  async () => {
-        sock.close()
-            fetch(props.url + '/logout', {
-            method: 'DELETE',
-            headers: {
-                'Accept': 'application/json',
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({'name': user.name, 'number': user.number})
-        })
-            .then(response => response.json())
-            .then(data => {
-                setUser({name:'', number:0})
-                navigate("/login")
+        if(leaveFromQueue.name === user.name) {
+            console.log(leaveFromQueue.name)
+            console.log(user.name)
+            setUser(oldUser => {
+                return {name: oldUser.name, number: 0}
             })
+        }
     }
 
     return (
         <div>
             <div>User: {user.name}</div>
-            <div>
-                <button disabled={!(user.number === 0)} onClick={takeQueue}>Take the queue</button>
-                <button disabled={(user.number === 0)} onClick={leaveQueue}>Leave the queue</button>
-                <button onClick={logOut}>Logout</button>
-            </div>
-            <div>
-                {queue.map((item) =>
-                    <Item item={item}/>
-                )}
-            </div>
+            <Navbar user={user} stompClient={stompClient}/>
+            <ItemList queue={queue}/>
         </div>
     );
 };
